feat : 카테고리 상품 등록 리펙토링

+++ tenshop-api/src/main/java/tenshop/api/product/dto/CategoryRegisterParam.java
@@ -2,6 +2,6 @@ package tenshop.api.product.dto;
 
 public record CategoryRegisterParam (
 	String name,
-	Integer parentId
+	Long parentId
 ) {
 }
